requirements/base.txt
Django==1.6.4
dj-database-url==0.3.0
django-choices==1.1.12
django-geojson==2.4.0
django-sortedm2m==0.7.0
django-storages==1.1.8
Pillow==2.4.0
South==0.8.4
whitenoise==1.0.2
